refactor Task components into our predefined folder structure format
